:recycle: Class 분리 작업 진행
