+++ src/main/java/kr/ac/hs/selab/auth/jwt/JwtSecurityConfig.java
@@ -1,11 +1,12 @@
 package kr.ac.hs.selab.auth.jwt;
 
-import javax.validation.constraints.NotNull;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.config.annotation.SecurityConfigurerAdapter;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.web.DefaultSecurityFilterChain;
 import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
+
+import javax.validation.constraints.NotNull;
 
 @RequiredArgsConstructor
 public class JwtSecurityConfig extends
