+++ src/main/java/kr/ac/hs/selab/auth/jwt/JwtTokenProvider.java
@@ -17,7 +17,6 @@ import org.springframework.stereotype.Component;
 import java.security.Key;
 import java.util.Arrays;
 import java.util.Collection;
-import java.util.Date;
 import java.util.stream.Collectors;
 
 // TODO : 조금씩 뿌시자!
@@ -27,11 +26,8 @@ import java.util.stream.Collectors;
 public class JwtTokenProvider implements InitializingBean {
 
     private static final String AUTHORITIES_KEY = "auth";
-    private static final int TOKEN_VALIDITY_TIME = 1000;
-    private static final String CLAIMS_REGEX = ",";
     private static final String EMPTY_REGEX = "";
     private final JwtProperties jwtProperties;
-
     private Key key;
 
     @Override
@@ -43,33 +39,12 @@ public class JwtTokenProvider implements InitializingBean {
         return Decoders.BASE64.decode(jwtProperties.getSecret());
     }
 
-    public String createToken(final Authentication authentication) {
-        var authorities = authoritiesToString(authentication);
-        var now = currentTime();
-        var validity = expireTime(now);
-
-        return Jwts.builder()
-                .setSubject(authentication.getName())
-                .setIssuer(jwtProperties.getIssuer())
-                .claim(AUTHORITIES_KEY, authorities)
-                .signWith(key, SignatureAlgorithm.HS512)
-                .setExpiration(validity)
-                .compact();
-    }
-
-    private String authoritiesToString(Authentication authentication) {
-        return authentication.getAuthorities().stream()
-                .map(GrantedAuthority::getAuthority)
-                .collect(Collectors.joining(CLAIMS_REGEX));
-    }
-
-    private long currentTime() {
-        return (new Date()).getTime();
-    }
-
-    private Date expireTime(long now) {
-        final var time = jwtProperties.getTokenValidityInSeconds() * TOKEN_VALIDITY_TIME;
-        return new Date(now + time);
+    public String create(Authentication authentication) {
+        return new Jwt(
+                jwtProperties.getIssuer(),
+                key,
+                jwtProperties.getTokenValidityInSeconds()
+        ).createToken(authentication);
     }
 
     // TODO : Provider와 상관 없는 내용임으로 분리해야 한다.
@@ -94,7 +69,7 @@ public class JwtTokenProvider implements InitializingBean {
 
     // 이건아님
     private Collection<? extends GrantedAuthority> makeAuthorities(Claims claims) {
-        return Arrays.stream(claims.get(AUTHORITIES_KEY).toString().split(CLAIMS_REGEX))
+        return Arrays.stream(claims.get(AUTHORITIES_KEY).toString().split(","))
                 .map(SimpleGrantedAuthority::new)
                 .collect(Collectors.toList());
     }
