+++ src/main/java/kr/ac/hs/selab/common/config/SecurityConfig.java
@@ -4,8 +4,6 @@ import kr.ac.hs.selab.auth.jwt.JwtAccessDeniedHandler;
 import kr.ac.hs.selab.auth.jwt.JwtAuthenticationEntryPoint;
 import kr.ac.hs.selab.auth.jwt.JwtSecurityConfig;
 import kr.ac.hs.selab.auth.jwt.JwtTokenProvider;
-import kr.ac.hs.selab.common.config.CorsConfig;
-import kr.ac.hs.selab.common.config.SwaggerConfig;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
